Изменение концепта передачи данных Areas.
