+++ app/src/main/java/ru/practicum/android/diploma/domain/repositories/AreasRepository.kt
@@ -1,7 +1,5 @@
 package ru.practicum.android.diploma.domain.repositories
 
-import ru.practicum.android.diploma.domain.models.Areas
-
 interface AreasRepository {
-    suspend fun getAreas(): Result<List<Areas>>
+    suspend fun fetchAreas()
 }
